+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ v1.2
 - Support packet capture and tracing for UAVCAN/CAN.
   Refactor the unstable capture/tracing API to model the underlying protocols more accurately.
 
+- Constructor parameter ``anonymous`` for ``UDPTransport`` has been deprecated in favor of ``local_node_id``.
+
 v1.1
 ----